update readme.md
Pivoting product vision description to remove any potential future features that had not been built or demo'd yet such as location data tracking collection or storing data. Our product descriptions showed possible product features or vision, but we still hadn't validate. 
Had a conversation today with with Harper Reed, who is in contact with epidemiological experts who verified some questions we have had with no answers as of yet. Today we heard that epidemiologists are asking for technology and apps to notify users of contact with someone who tested positive, but that going farther with data collection is not actually useful and they are NOT asking for that. still need to talk to epidemiologist directly but hearing that more data collection is not useful to them since the way they track the epidemic is through 1. a positive test, then 2. through interviews assess who else has had contact. now that there is a general privacy consensus from the health organizations as well as the data and tech communities we can align with these. For employers a recommendation that is in alignment with not overstepping privacy concerns stops at contact tracing and providing support for employees through testing access and coverage to medical resources. Notifying users of contact and then having them get tested and follow health care org's recommendations.
Currently demo only utilizes contact tracing. We removed admin visibility from our roadmap, and continue to work on optimizing BLE code that Aaron has been working on prototyping and improving for standard contact tracing.